Fixes broken compilation under arduino ide 1.6.5-r5
"alarm_clock:38: error: 'alarm_t' was not declared in this scope #2"

In the pre-compilation mangling the IDE does to turn the .ino file into a .cpp file it was inserting (unecessary) function prototypes which was causing my first line of real code (a typedef) to be an error. Moving this into a header resolves the issue.

This was not present in the 'old' version of the IDE that comes with Debian Jessie (1.0.5) or building with the Makefile in the Debian arduino-mk package